+++ src/admin/class-admin-ajax.php
@@ -26,18 +26,7 @@ class Admin_Ajax {
 	 */
 	public static function register() {
 		/* Generic AJAX Handlers */
-		add_action( 'wp_ajax_ptc_get_post_options_by_title', [ __CLASS__, 'ajax_ptc_get_post_options_by_title' ] );
 		add_action( 'wp_ajax_ptc_get_post_title_by_id', [ __CLASS__, 'ajax_ptc_get_post_title_by_id' ] );
-	}
-
-	/**
-	 * AJAX handler to get post options by like title.
-	 *
-	 * @since 3.0.0 Moved to Admin_Ajax class.
-	 * @since 1.1.0
-	 */
-	public static function ajax_ptc_get_post_options_by_title() {
-		require_once PLUGIN_PATH . 'src/admin/ajax/ajax-get-post-options-by-title.php';
 	}
 
 	/**