+++ src/includes/class-autoloader.php
@@ -78,6 +78,7 @@ class Autoloader {
 		'REST_API\Tasks'                   => 'src/public/rest-api/class-tasks.php',
 		'REST_API\Automations'             => 'src/public/rest-api/class-automations.php',
 		'REST_API\Tags'                    => 'src/public/rest-api/class-tags.php',
+		'REST_API\Posts'                   => 'src/public/rest-api/class-posts.php',
 
 	);